+++ crismaapp/__init__.py
@@ -1,1 +1,5 @@
-from .routes import app
+from . import crismandos
+from . import domingos
+from . import encontros
+
+from .app import app
